requirements.txt
django
django-htmx
google-genai
python-dotenv
markdown
boto3
